Add Operators table
Operators is a table (list of lists) of operator precedences. You can
query it by running Operators.table. Now, the language (and the user in
the future) will be able to change order of messages based on the table.

The eventual goal is to allow user to add operators as he goes along,
however that would require a decently large change to how we parse
statements. A more closer goal is implementing unary operators such as
-, ! and return.

This commit also includes code cleanup in context.cpp (should be cleaned
up further in the future), a bug fix in parse_message_tail and
implementation of two operands, + and * as I wanted to write a test to
check if precedence swap worked correctly.