Implement this accessor in functions
this refers to the object that the function is part of. The syntax
however stays the same, e.g. to modify a store of the "external" object,
you write this store = ...;

Additionally, fix a parsing bug with an if statement nested inside a
function. Because of the trailing } brackets, parse_statement_list()
would overparse everything following the function definition as
additional statements in the function body.