Start implementing bytecode interpretation
Before this, we were just traversing the AST and executing as we go
along. However, this got really cumbersome and unpleasant to work with
(e.g. the gigantic block of if/else for default stores, no well-defined scope
needed for Callable etc.). Bytecode should make this well easier.

So far, the bytecode has three instructions, Load (load value into
register), Send (send message to a r1 and store result in r2) and
Store (make a new store in r1 and store contents of r2 in the store).
Surpisingly, these instructions should be enough to interpret everything
that we had before.

Next, extra code generation should be removed from parser (e.g.
how it parses Int, Char, String) and after that we can implement the
basic interpreter.