Add stamps to messages, move basic object cloning to generate_bytecode
Stamps are what was before called sender - its either a value or an
Object. Its represented by optional<variant<Register, string>>.

Basic objects are (so far) Int, Char and String. The macros written
should make it easy to work with them in the future. Ultimately, the
bytecode generated is the same as if the parser had not been modified,
however now it is handled by the bytecode generation, which makes more
sense.