Add bytecode generation for if statement
For this, the parsing of an if statement changed since we no longer do
if as a message to an if object. Now it is handled via a bytecode
instruction - JumpFalse(register, destination_basic_block). A pair
instruction, JumpTrue was added. It might be useful at the future, but
even if not we have it for completeness.