Add bytecode generation for function creation and call
Again, this is mostly just offloading work from the parser to the
bytecode generator. Additionally, basic blocks in the generator are now
stored on the heap. Previously, because they were stored next to each
other and instructions are cast to instruction, adding another block
could clobber the end of the last instruction in the previous basic
block.