Add ==, != to Object
We now have to pass in the interpreter to Object::send for working with
contexts.

True and False are stored in prelude.stamp. I do not know how we will
include the prelude - the easiest way would be to hard-code it, but we
might implement a basic "using" language construct for it.