Fully implemet function calls
This includes the new defaults of Callable: pass_param and call.
For this, we introduce a new bytecode instruction: JumpSaved. When a
call default is executed, it stores the return basic block, which
JumpSaved jumps to.
Additionally, string now stores the full string instead a Vec of Chars.
This makes more sence right now and makes String easier to operate on.