Add mut keyword
mut can only be applied to store statements, i.e. mut Dog barks = True;
It is possible to make a mutable store immutable (but not the other way
around), if mut is not applied every time a store is modified. This
makes sense to me: as this provides an easy way to change mutability of
a store. However, I recognize that this might make the syntax terse
during use, so this property is subject to change.