prelude.stamp is now automatically included before execution
What is actually included is the bytecode prelude.ostamp since it is
supposed to be faster. Because of this, new job was added to Makefile
that compiles prelude. It is run automatically if you make a release
version. Path is not resolved for the prelude file right now, which we
will have to deal with later.