Add return statement support
So far, return is a keyword - although the plan was to make it an
operator instead. However, since there is no support for operator
precedence yet, this is one way to implement it.

JumpSaved now stores the return register within the interpreter.

Additionally, fix multiple bugs related to function calling, jumping to
basic blocks, etc.